can set options through object wrapper
